Refactor: Extracted class from JmdslConfig
