+++ jmeter-java-dsl-cli/src/main/java/us/abstracta/jmeter/javadsl/cli/JmdslConfig.java
@@ -14,11 +14,11 @@ import picocli.CommandLine.ArgGroup;
 import picocli.CommandLine.Option;
 import picocli.CommandLine.Parameters;
 
-public class JmdslConfig {
+public class JmdslConfig extends JmdslApplyDefaults {
 
   public static final String CONFIG_OPTION = "--config";
   public static final String DEFAULT_CONFIG_FILE = ".jmdsl.yml";
-  private static final String PICOCLI_NO_DEFAULT_VALUE_MARKER = "__no_default_value__";
+  protected static final String PICOCLI_NO_DEFAULT_VALUE_MARKER = "__no_default_value__";
 
   private RecorderCommand recorder;
 
@@ -44,50 +44,6 @@ public class JmdslConfig {
       return;
     }
     applyDefaultsFromTo(other.recorder, this.recorder);
-  }
-
-  private static void applyDefaultsFromTo(Object defaults, Object target) {
-    Arrays.stream(target.getClass().getDeclaredFields())
-        .filter(f -> !f.isAnnotationPresent(JsonIgnore.class)
-            && (f.isAnnotationPresent(Option.class) || f.isAnnotationPresent(Parameters.class)))
-        .filter(f -> {
-          Object prevVal = getField(f, target);
-          String defaultValue = f.isAnnotationPresent(Option.class)
-              ? f.getAnnotation(Option.class).defaultValue()
-              : f.getAnnotation(Parameters.class).defaultValue();
-          if (PICOCLI_NO_DEFAULT_VALUE_MARKER.equals(defaultValue)) {
-            defaultValue = null;
-          }
-          return prevVal == null
-              || defaultValue != null && defaultValue.equals(prevVal.toString())
-              || prevVal instanceof Boolean && !(Boolean) prevVal
-              || prevVal instanceof List && (((List) prevVal).isEmpty());
-        })
-        .forEach(f -> setField(f, target, getField(f, defaults)));
-    Arrays.stream(target.getClass().getDeclaredFields())
-        .filter(f -> f.isAnnotationPresent(ArgGroup.class))
-        .forEach(f -> applyDefaultsFromTo(getField(f, defaults), getField(f, target)));
-  }
-
-  private static Object getField(Field field, Object instance) {
-    try {
-      field.setAccessible(true);
-      return field.get(instance);
-    } catch (IllegalAccessException e) {
-      throw new RuntimeException(e);
-    }
-  }
-
-  private static void setField(Field field, Object instance, Object value) {
-    // keep with default field value if value to set is null
-    if (value == null) {
-      return;
-    }
-    try {
-      field.set(instance, value);
-    } catch (IllegalAccessException e) {
-      throw new RuntimeException(e);
-    }
   }
 
 }
